# compiler: zero-copy code assembly