--- a/internal/asm/amd64/impl_staticconst.go
+++ b/internal/asm/amd64/impl_staticconst.go
@@ -12,7 +12,7 @@ import (
 // limit plus max(length(c) for c in the pool) so we must ensure that limit is less than 2^31.
 const defaultMaxDisplacementForConstantPool = 1 << 30
 
-func (a *AssemblerImpl) maybeFlushConstants(isEndOfFunction bool) {
+func (a *AssemblerImpl) maybeFlushConstants(buf asm.Buffer, isEndOfFunction bool) {
 	if a.pool.Empty() {
 		return
 	}
@@ -21,7 +21,7 @@ func (a *AssemblerImpl) maybeFlushConstants(isEndOfFunction bool) {
 		// If the distance between (the first use in binary) and (end of constant pool) can be larger
 		// than MaxDisplacementForConstantPool, we have to emit the constant pool now, otherwise
 		// a const might be unreachable by a literal move whose maximum offset is +- 2^31.
-		((a.pool.PoolSizeInBytes+a.buf.Len())-int(a.pool.FirstUseOffsetInBinary)) >= a.MaxDisplacementForConstantPool {
+		((a.pool.PoolSizeInBytes+buf.Len())-int(a.pool.FirstUseOffsetInBinary)) >= a.MaxDisplacementForConstantPool {
 		if !isEndOfFunction {
 			// Adds the jump instruction to skip the constants if this is not the end of function.
 			//
@@ -29,25 +29,24 @@ func (a *AssemblerImpl) maybeFlushConstants(isEndOfFunction bool) {
 			// small enough to fit all consts after the end of function.
 			if a.pool.PoolSizeInBytes >= math.MaxInt8-2 {
 				// long (near-relative) jump: https://www.felixcloutier.com/x86/jmp
-				a.buf.WriteByte(0xe9)
-				a.writeConst(int64(a.pool.PoolSizeInBytes), 32)
+				buf.WriteByte(0xe9)
+				buf.WriteUint32(uint32(a.pool.PoolSizeInBytes))
 			} else {
 				// short jump: https://www.felixcloutier.com/x86/jmp
-				a.buf.WriteByte(0xeb)
-				a.writeConst(int64(a.pool.PoolSizeInBytes), 8)
+				buf.Write2Bytes(0xeb, byte(a.pool.PoolSizeInBytes))
 			}
 		}
 
 		for _, c := range a.pool.Consts {
-			c.SetOffsetInBinary(uint64(a.buf.Len()))
-			a.buf.Write(c.Raw)
+			c.SetOffsetInBinary(uint64(buf.Len()))
+			buf.Write(c.Raw)
 		}
 
 		a.pool.Reset()
 	}
 }
 
-func (a *AssemblerImpl) encodeRegisterToStaticConst(n *nodeImpl) (err error) {
+func (a *AssemblerImpl) encodeRegisterToStaticConst(buf asm.Buffer, n *nodeImpl) (err error) {
 	var opc []byte
 	var rex byte
 	switch n.instruction {
@@ -58,7 +57,7 @@ func (a *AssemblerImpl) encodeRegisterToStaticConst(n *nodeImpl) (err error) {
 	default:
 		return errorEncodingUnsupported(n)
 	}
-	return a.encodeStaticConstImpl(n, opc, rex, 0)
+	return a.encodeStaticConstImpl(buf, n, opc, rex, 0)
 }
 
 var staticConstToRegisterOpcodes = [...]struct {
@@ -91,7 +90,7 @@ var staticConstToRegisterOpcodes = [...]struct {
 	ADDQ: {opcode: []byte{0x03}, rex: rexPrefixW},
 }
 
-func (a *AssemblerImpl) encodeStaticConstToRegister(n *nodeImpl) (err error) {
+func (a *AssemblerImpl) encodeStaticConstToRegister(buf asm.Buffer, n *nodeImpl) (err error) {
 	var opc []byte
 	var rex, mandatoryPrefix byte
 	info := staticConstToRegisterOpcodes[n.instruction]
@@ -105,13 +104,13 @@ func (a *AssemblerImpl) encodeStaticConstToRegister(n *nodeImpl) (err error) {
 	default:
 		opc, rex, mandatoryPrefix = info.opcode, info.rex, info.mandatoryPrefix
 	}
-	return a.encodeStaticConstImpl(n, opc, rex, mandatoryPrefix)
+	return a.encodeStaticConstImpl(buf, n, opc, rex, mandatoryPrefix)
 }
 
 // encodeStaticConstImpl encodes an instruction where mod:r/m points to the memory location of the static constant n.staticConst,
 // and the other operand is the register given at n.srcReg or n.dstReg.
-func (a *AssemblerImpl) encodeStaticConstImpl(n *nodeImpl, opcode []byte, rex rexPrefix, mandatoryPrefix byte) (err error) {
-	a.pool.AddConst(n.staticConst, uint64(a.buf.Len()))
+func (a *AssemblerImpl) encodeStaticConstImpl(buf asm.Buffer, n *nodeImpl, opcode []byte, rex rexPrefix, mandatoryPrefix byte) (err error) {
+	a.pool.AddConst(n.staticConst, uint64(buf.Len()))
 
 	var reg asm.Register
 	if n.dstReg != asm.NilRegister {
@@ -125,28 +124,26 @@ func (a *AssemblerImpl) encodeStaticConstImpl(n *nodeImpl, opcode []byte, rex re
 
 	var instLen int
 	if mandatoryPrefix != 0 {
-		a.buf.WriteByte(mandatoryPrefix)
+		buf.WriteByte(mandatoryPrefix)
 		instLen++
 	}
 
 	if rexPrefix != rexPrefixNone {
-		a.buf.WriteByte(rexPrefix)
+		buf.WriteByte(rexPrefix)
 		instLen++
 	}
 
-	a.buf.Write(opcode)
+	buf.Write(opcode)
 	instLen += len(opcode)
 
 	// https://wiki.osdev.org/X86-64_Instruction_Encoding#32.2F64-bit_addressing
 	modRM := 0b00_000_101 | // Indicate "[RIP + 32bit displacement]" encoding.
 		(reg3Bits << 3) // Place the reg on ModRM:reg.
-	a.buf.WriteByte(modRM)
+	buf.WriteByte(modRM)
 	instLen++
 
 	// Preserve 4 bytes for displacement which will be filled after we finalize the location.
-	for i := 0; i < 4; i++ {
-		a.buf.WriteByte(0)
-	}
+	buf.Write4Bytes(0, 0, 0, 0)
 	instLen += 4
 
 	if !n.staticConstReferrersAdded {