--- a/internal/engine/compiler/compiler.go
+++ b/internal/engine/compiler/compiler.go
@@ -16,9 +16,9 @@ type compiler interface {
 	// compilePreamble is called before compiling any wazeroir operation.
 	// This is used, for example, to initialize the reserved registers, etc.
 	compilePreamble() error
-	// compile generates the byte slice of native code.
+	// compile generates the native code into buf.
 	// stackPointerCeil is the max stack pointer that the target function would reach.
-	compile() (code []byte, stackPointerCeil uint64, err error)
+	compile(buf asm.Buffer) (stackPointerCeil uint64, err error)
 	// compileGoHostFunction adds the trampoline code from which native code can jump into the Go-defined host function.
 	// TODO: maybe we wouldn't need to have trampoline for host functions.
 	compileGoDefinedHostFunction() error