This PR modifies the assembler and compiler to remove the copy from the assembler's internal buffer to the memory-mapped code segments.

To do this, I introduced two types in `internal/asm`: `CodeSegment` and `Buffer`. The former represents the code segments generated by the compiler, and the latter is similar to `bytes.Buffer` but is backed by a `CodeSegment` to support assembling instructions directly into the memory-mapped regions.

The intent is to further reduce the memory footprint of the compiler by dropping the need for the intermediary buffer but also to improve the compute time by avoiding a copy and creating opportunities to write instructions more efficiently.

Regarding the second point, I added two methods on the `Buffer` type to improve the assembler throughput.

On ARM64, all instructions are 4 bytes long, so instead of using a `Write([]byte)` method (which always makes a call to `runtime.memmove`), I added a `Write4Bytes` methods which combines all four bytes into a 32 bits integer and writes it to the underlying buffer. 

On AMD64, I revisited the implementation to support writing directly to the output buffer via calls to Go's `append` builtin. This allows the compiler to generate much better code, almost always inlining the writes.

## Benchmarks

The first round of benchmarks I am recording here comes from the internal integration tests. In those benchmarks, we see observe a measurable drop in memory utilization (5-14%) and a 10% lower compute footprint on amd64. 
```
$ go test -v -run '^$' -bench BenchmarkCompilation -benchmem -count 6 ./internal/integration_test/bench/
```
```
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
                                 │ /tmp/bench.0 │            /tmp/bench.1            │
                                 │    sec/op    │   sec/op     vs base               │
Compilation/with_extern_cache       317.0µ ± 3%   316.7µ ± 2%        ~ (p=0.937 n=6)
Compilation/without_extern_cache    5.659m ± 2%   5.056m ± 2%  -10.65% (p=0.002 n=6)
Compilation/interpreter             1.257m ± 0%   1.247m ± 1%   -0.80% (p=0.002 n=6)
geomean                             1.311m        1.259m        -3.97%

                                 │ /tmp/bench.0 │            /tmp/bench.1            │
                                 │     B/op     │     B/op      vs base              │
Compilation/with_extern_cache      48.96Ki ± 0%   48.94Ki ± 0%  -0.04% (p=0.002 n=6)
Compilation/without_extern_cache   629.9Ki ± 0%   597.9Ki ± 0%  -5.08% (p=0.002 n=6)
Compilation/interpreter            759.0Ki ± 0%   759.0Ki ± 0%       ~ (p=0.327 n=6)
geomean                            286.0Ki        281.1Ki       -1.74%

                                 │ /tmp/bench.0 │            /tmp/bench.1             │
                                 │  allocs/op   │  allocs/op   vs base                │
Compilation/with_extern_cache        420.0 ± 0%    420.0 ± 0%       ~ (p=1.000 n=6) ¹
Compilation/without_extern_cache    1.083k ± 0%   1.073k ± 0%  -0.88% (p=0.002 n=6)
Compilation/interpreter              627.0 ± 0%    626.5 ± 0%       ~ (p=0.545 n=6)
geomean                              658.2         656.1       -0.32%
¹ all samples are equal
```
```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
                                 │ /tmp/bench.0 │           /tmp/bench.1            │
                                 │    sec/op    │   sec/op     vs base              │
Compilation/with_extern_cache       148.4µ ± 2%   149.6µ ± 3%       ~ (p=0.818 n=6)
Compilation/without_extern_cache    1.902m ± 6%   1.851m ± 5%       ~ (p=0.093 n=6)
Compilation/interpreter             663.7µ ± 1%   658.0µ ± 1%  -0.86% (p=0.015 n=6)
geomean                             572.3µ        566.9µ       -0.93%

                                 │ /tmp/bench.0 │            /tmp/bench.1             │
                                 │     B/op     │     B/op      vs base               │
Compilation/with_extern_cache      48.81Ki ± 0%   48.80Ki ± 0%        ~ (p=0.374 n=6)
Compilation/without_extern_cache   655.5Ki ± 0%   565.7Ki ± 0%  -13.71% (p=0.002 n=6)
Compilation/interpreter            758.9Ki ± 0%   758.9Ki ± 0%        ~ (p=0.548 n=6)
geomean                            289.6Ki        275.7Ki        -4.80%

                                 │ /tmp/bench.0 │            /tmp/bench.1            │
                                 │  allocs/op   │ allocs/op   vs base                │
Compilation/with_extern_cache        420.0 ± 0%   420.0 ± 0%       ~ (p=1.000 n=6) ¹
Compilation/without_extern_cache     938.0 ± 0%   934.0 ± 0%  -0.43% (p=0.002 n=6)
Compilation/interpreter              626.0 ± 0%   626.0 ± 0%       ~ (p=1.000 n=6) ¹
geomean                              627.1        626.2       -0.14%
¹ all samples are equal
```

The second set of benchmarks here show the differences on the compilation of Python, which I usually run this way:
```sh
# run a few passes of compilation to measure both the baseline and the change
$ time ./wazero compile -count 5 -cpuprofile /tmp/profile.out ~/wasm/bin/python-3.11.3.wasm
...
# analyse with pprof
$ go tool pprof -text -base /tmp/profile.base /tmp/profile.out
```

Overall, the most interesting gains here are a 7% drop in compilation time on amd64, and a few MiB shaved off by removing the intermediary buffer. We're not seeing similar improvements on CPU usage on arm64 tho I'm wondering if this is due to Darwin not dealing with page faults on memory mappings as well as Linux, it seems a lot more time is spent writing to those memory segments (so it might be the bottleneck).

**linux/amd64**
```
File: wazero
Type: cpu
Time: May 18, 2023 at 12:14am (PDT)
Duration: 15.95s, Total samples = 15.72s (98.58%)
Showing nodes accounting for -1.10s, 7.00% of 15.72s total
Dropped 10 nodes (cum <= 0.08s)
      flat  flat%   sum%        cum   cum%
     0.79s  5.03%  5.03%     -0.64s  4.07%  github.com/tetratelabs/wazero/internal/asm/amd64.(*AssemblerImpl).encode
    -0.58s  3.69%  1.34%     -1.17s  7.44%  github.com/tetratelabs/wazero/internal/asm/amd64.(*AssemblerImpl).maybeNOPPadding
    -0.56s  3.56%  2.23%     -0.85s  5.41%  bytes.(*Buffer).WriteByte
```
```
File: wazero
Type: alloc_space
Time: May 18, 2023 at 12:13am (PDT)
Showing nodes accounting for -15.31MB, 1.67% of 917.83MB total
Dropped 3 nodes (cum <= 4.59MB)
      flat  flat%   sum%        cum   cum%
  -12.04MB  1.31%  1.31%   -12.04MB  1.31%  github.com/tetratelabs/wazero/internal/engine/compiler.(*runtimeValueLocationStack).push (inline)
   -9.80MB  1.07%  2.38%    -9.80MB  1.07%  bytes.growSlice
    5.73MB  0.62%  1.76%     6.73MB  0.73%  github.com/tetratelabs/wazero/internal/wasm/binary.decodeDataSegment
   -5.55MB  0.61%  2.36%    -5.55MB  0.61%  github.com/tetratelabs/wazero/internal/asm/amd64.(*nodePool).allocNode (inline)
```

**darwin/arm64**
```
Type: cpu
Time: May 18, 2023 at 12:03am (PDT)
Duration: 5.69s, Total samples = 5.31s (93.39%)
Showing nodes accounting for -0.13s, 2.45% of 5.31s total
Dropped 8 nodes (cum <= 0.03s)
      flat  flat%   sum%        cum   cum%
     1.61s 30.32% 30.32%      1.61s 30.32%  encoding/binary.littleEndian.PutUint32 (inline)
    -1.44s 27.12%  3.20%     -1.44s 27.12%  runtime.memmove
    -0.06s  1.13%  2.07%     -0.16s  3.01%  bytes.(*Buffer).Write
    -0.06s  1.13%  0.94%      0.60s 11.30%  github.com/tetratelabs/wazero/internal/asm/arm64.(*AssemblerImpl).encodeLoadOrStoreWithConstOffset
```
```
Type: alloc_space
Time: May 18, 2023 at 12:06am (PDT)
Showing nodes accounting for -45.46MB, 5.12% of 887.49MB total
Dropped 3 nodes (cum <= 4.44MB)
      flat  flat%   sum%        cum   cum%
  -22.27MB  2.51%  2.51%   -22.27MB  2.51%  bytes.growSlice
  -16.13MB  1.82%  4.33%   -16.13MB  1.82%  github.com/tetratelabs/wazero/internal/asm/arm64.(*nodePool).allocNode (inline)
   -5.50MB  0.62%  4.95%    -5.50MB  0.62%  debug/dwarf.(*Data).parseAbbrev
    4.95MB  0.56%  4.39%     4.95MB  0.56%  github.com/tetratelabs/wazero/internal/bitpack.newDeltaArray[...] (inline)
   -4.03MB  0.45%  4.84%    -4.03MB  0.45%  github.com/tetratelabs/wazero/internal/engine/compiler.(*runtimeValueLocationStack).push (inline)
```

Finally, I ran `make fuzz fuzz_timeout_seconds=180` to increase confidence in the change since it is touching a lot of core pieces of the compiler:
```
Done 358310 runs in 181 second(s)
```

## Comments

Overall, this is a lot of code change for a ~10% improvement, which is likely a sign that we are entering the land of diminishing returns in this direction. I wanted to measure how much the removal of the intermediary assembler buffer would impact, so now we know! But I've been too close to the issue to emit a fair judgment on whether the gains are worth the change.

Feel free to chime in with any comments or feedback!